fix: task results mq listener coroutine cancelled error (#15)
* feat: custom async consumer with `pika`

Tested it by dispatching it on the `uvloop` event loop created by
`fastapi`. Works well but it only acks and logs the message, does not do
any proper message processing now.

* refactor: use `AsyncConsumer` for task results processing

* chore: remove `aio-pika` as dependency

* refactor: change to topic ampq exchange

This allows for better flexibility when it comes to monitoring. All
dispatchers (aka backend) and worker goes through the same exchange and
the task and resuilt queues are binded to that single exchange