feat: update task (#64)
* refactor(mcq/mrq): add choice class

Choices were represented as list[str] with the answer as the index.
By separating the user's answer (id) from the display order
(order_index), past attempts will be less vulnerable to task updates.

* feat(task): add updated_version_id + order_index

- migration for these two columns
- also update json definitions to add these 2 columns
- also sort submission task attempts by these columns

* fix(problem): order tasks by order_index

* fix: revert order_by fix

breaks the query with
```
 ProgrammingError: (psycopg2.errors.UndefinedTable) missing
         FROM-clause entry for table "task"
         LINE 3: WHERE submission_1.id IN (5, 4, 3, 2, 1) ORDER BY
         task.order...
```
so change reverted - but task attempts in submissions
possibly not in correct order

* chore: regenerate migrations

* feat(task): init update endpoint

* feat(task): add delete endpoint

- also modify task -> task_attempt -> task_result to use cascading
  delete

* fix(task): block changes to "soft deleted" tasks

return 404 error for update/delete task if updated_version_id is not
None

* feat(task): implement reorder

* feat(task): add rerun endpoint

* fix(task): update order index on addition

* refactor(choice): use string (uuid) for id

why str and not uuid type = sqlalchemy will give not serializable error

* fix(submission.json): use correct uuids

* fix: types

* fix: types

* fix(task): rerun correct task attempt on update